+++ sphinx-docs/source/changelog.rst
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+v2.3 (2021-03-11)
+--------------------
+
+*New Features*
+
+- Can generate HOOMD snapshot bonds group based on mapped exclusion lists
+
+
 v2.21 (2021-03-08)
 
 *Bug Fixes*
@@ -15,7 +23,6 @@ v2.2 (2021-03-01)
 - Can read molecules from GROAMCS topology
 - Can compute one-hot encoding for different bead type interactions in CG systems
 - Can generate mapped exclusion lists in CG systems
-- Can generate HOOMD snapshot bonds group based on mapped exclusion lists
 
 *Enhancements*